+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v2.0.0: 2016-12-28
+
+- Use [extend-shallow](https://www.npmjs.com/package/extend-shallow) to merge the data
+- Adds `metadata-files` property to allow inheriting metadata information
+
 ## v1.0.1: 2016-12-28
 
 - Use [soft-extend](http://npm.im/soft-extend) to merge the data